ASSERTION FAILED: m_truncation != cFullTruncation in InlineTextBox::clampedOffset() https://bugs.webkit.org/show_bug.cgi?id=178322
Reviewed by Darin Adler.

Source/WebCore:

It is acceptable for InlineTextBox::clampedOffset() to be called for a fully truncated box,
say a person clicks on the ellipsis in a truncated text run. Restore the behavior prior to
r223259 and return the clamped offset.

Test: fast/text/click-ellipsis-assertion-failure.html

* rendering/InlineTextBox.cpp:
(WebCore::InlineTextBox::clampedOffset const):

LayoutTests:

Add a test to ensure that clicking on the ellipsis in a truncated text run does not
cause an assertion failure in a debug build.

* fast/text/click-ellipsis-assertion-failure-expected.txt: Added.
* fast/text/click-ellipsis-assertion-failure.html: Added.

Identifier: 194615@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223553 268f45cc-cd09-0410-ab3c-d52691b4dbfc