Turn on ccache for Mac cmake builds by default https://bugs.webkit.org/show_bug.cgi?id=177059
Reviewed by Sam Weinig.

.:

* Source/cmake/WebKitCCache.cmake: Added.
* Source/cmake/WebKitCommon.cmake:
Turn on ccache for Mac CMake builds (Makefile and Ninja generators only)
if it's installed, making use of CMake's ability to wrap the compiler invocation.

Tools:

* ccache/ccache-clang:
* ccache/ccache-wrapper: Added.
Add a pass-through ccache wrapper to be used with CMake, in addition
to the existing faux-clang wrappers.

* Scripts/build-webkit:
Add --use-ccache and --no-use-ccache option, which will define
WK_USE_CCACHE to YES or NO, respectively, which the underlying
build systems respect. We do not define WK_USE_CCACHE if the option
is not specified, because the underlying build systems have different
default values.


Identifier: 194816@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223818 268f45cc-cd09-0410-ab3c-d52691b4dbfc