[GTK][WPE] generate-forwarding-headers.pl should also scan derived sources in WebKit framework https://bugs.webkit.org/show_bug.cgi?id=178579
Reviewed by Carlos Garcia Campos.

* PlatformGTK.cmake: When generating forwarding headers, derived sources
should be scanned as well. The DERIVED_SOURCES_WEBKIT_DIR value has to
be passed as an include path to the generate-forward-headers.pl script
to achieve that. This target then also has to depend on all the derived
source files to be generated before this can be performed.
* PlatformWPE.cmake: Ditto.


Identifier: 194880@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223884 268f45cc-cd09-0410-ab3c-d52691b4dbfc