[GTK][X11] Windy.com shows always straight wind lines https://bugs.webkit.org/show_bug.cgi?id=176718
Reviewed by Carlos Garcia Campos.

WebGL's GL_LUMINANCE_ALPHA format is not available in OpenGL when using a version >= 3.2
and a core profile. In that case, we need to replace it with GL_RG and swizzle the color
components appropriately.

No new behavior.

* platform/graphics/opengl/GraphicsContext3DOpenGL.cpp:
(WebCore::GraphicsContext3D::texImage2D):
* platform/graphics/opengl/GraphicsContext3DOpenGLCommon.cpp:
(WebCore::GraphicsContext3D::texSubImage2D):


Identifier: 194883@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223887 268f45cc-cd09-0410-ab3c-d52691b4dbfc